# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentChannelResponse:
    properties:
      isActive:
        description: 0 = no signal, 1 = has signal
        type: integer
      isPhoton:
        description: 0 = analog, 1 = photon
        type: integer
      polarization:
        type: string
      wavelen:
        type: number
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentChannelsResponse:
    properties:
      channels:
        items:
          $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentChannelResponse'
        type: array
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentResponse:
    properties:
      comments:
        type: string
      created_at:
        type: string
      error_msg:
        type: string
      id:
        type: integer
      licel_bgr_path:
        type: string
      licel_zip_path:
        type: string
      measurement_start_time:
        type: string
      measurement_stop_time:
        type: string
      meteo_file_path:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.GlueExperimentBody:
    properties:
      h1:
        type: number
      h2:
        type: number
      polarization:
        type: string
      wavelengths:
        items:
          type: number
        minItems: 1
        type: array
    required:
    - h1
    - h2
    - wavelengths
    type: object
...
  github_com_kshmirko_lidar-platform-go_pkg_dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.PrepareExperimentBody:
    properties:
      bgr_alt:
        minimum: 0
        type: number
      bgr_type:
        enum:
        - file
        - avgTail
        - medTail
        type: string
      crop_alt:
        minimum: 0
        type: number
    required:
    - bgr_type
    - crop_alt
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.PreparedExperimentResponse:
    properties:
      bgr_alt:
        type: number
      bgr_type:
        type: string
      crop_alt:
        type: number
      error_msg:
        type: string
      experiment_id:
        type: integer
      id:
        type: integer
      path_to_data:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
  github_com_kshmirko_lidar-platform-go_pkg_dto.TaskStatusResponse:
    properties:
      error:
        description: error message (when failed)
        type: string
      status:
        description: pending, processing, done, failed
        type: string
      task_id:
        type: string
      url:
        description: presigned Minio URL (when done)
        type: string
    type: object
...
  github_com_kshmirko_lidar-platform-go_pkg_dto.VisualizeChartResponse:
    properties:
      status:
        description: '"accepted" — task was enqueued'
        type: string
      task_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@lidar-platform.io
    name: API Support
  description: REST API for the Lidar Platform — user management, point cloud processing,
    and device telemetry.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lidar Platform API
  version: "1.0"
paths:
...
  /experiments:
    get:
      description: Returns a paginated list of experiments with optional filtering.
      parameters:
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Filter by status
        enum:
        - staged
        - uploading
        - done
        - failed
        in: query
        name: status
        type: string
      - description: Filter by title (case-insensitive partial match)
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentPaginatedResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all experiments
      tags:
      - experiments
    post:
      consumes:
      - multipart/form-data
      description: Uploads licel zip, bgr and meteo files, creates an experiment and
        starts asynchronous preprocessing.
      parameters:
      - description: Experiment title
        in: formData
        name: title
        required: true
        type: string
      - description: Experiment comments
        in: formData
        name: comments
        type: string
      - description: Licel measurements zip archive
        in: formData
        name: licelZip
        required: true
        type: file
      - description: Licel BGR file
        in: formData
        name: licelBgr
        required: true
        type: file
      - description: Meteo data file
        in: formData
        name: meteoFile
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create experiment
      tags:
      - experiments
  /experiments/{id}:
    get:
      description: Returns a single experiment by its database ID.
      parameters:
      - description: Experiment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get experiment by ID
      tags:
      - experiments
  /experiments/{id}/channels:
    get:
      description: Returns the list of available measurement channels (wavelength,
        polarization, photon/analog) for an experiment.
      parameters:
      - description: Experiment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ExperimentChannelsResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get experiment channels
      tags:
      - experiments
  /experiments/{id}/glue:
    post:
      consumes:
      - application/json
      description: Starts asynchronous channel gluing for specified wavelengths and
        altitude range.
      parameters:
      - description: Experiment ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Glue parameters: wavelengths, polarization, altitude range h1-h2'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.GlueExperimentBody'
      produces:
      - application/json
      responses:
        "202":
          description: Glue task submitted
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.MessageResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "409":
          description: Invalid experiment status
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Glue experiment channels
      tags:
      - experiments
  /experiments/{id}/prepare:
    post:
      consumes:
      - application/json
      description: 'Starts background processing: background subtraction and cropping.
        Stores result in Minio.'
      parameters:
      - description: Experiment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Preparation parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.PrepareExperimentBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.PreparedExperimentResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "404":
          description: Experiment not found
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "409":
          description: Experiment not ready
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Prepare experiment data
      tags:
      - experiments
  /prepared/{id}:
    get:
      description: Enqueues a chart generation task (heatmap or profile) and returns
        a task ID for polling. Poll /tasks/{taskID} for the result.
      parameters:
      - description: Prepared experiment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Wavelength
        format: float64
        in: query
        name: wavelen
        required: true
        type: number
      - default: 0
        description: 'Photon/analog mode: 0=analog (default), 1=photon; ignored when
          glued=1'
        in: query
        name: photon
        type: integer
      - default: o
        description: Polarization
        in: query
        name: polarization
        type: string
      - description: image or profile
        enum:
        - image
        - profile
        in: query
        name: action
        required: true
        type: string
      - default: 0
        description: 'Glued mode: 0=non-glued (default), 1=glued'
        enum:
        - 0
        - 1
        in: query
        name: glued
        type: integer
      - default: png
        description: 'Output type: png, svg or json'
        enum:
        - png
        - svg
        - json
        in: query
        name: type
        type: string
      - default: raw
        description: 'Signal formula: raw, rangecorr, lograngecorr'
        enum:
        - raw
        - rangecorr
        - lograngecorr
        in: query
        name: formula
        type: string
      - default: false
        description: Force regeneration, ignoring cache
        in: query
        name: regenerate
        type: boolean
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.VisualizeChartResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Visualize prepared experiment data (async)
      tags:
      - experiments
  /tasks/{taskID}:
    get:
      description: Returns the current status of an async task (pending, processing,
        done, failed). When done, includes the presigned chart URL.
      parameters:
      - description: Task ID returned by /prepared/{id}
        in: path
        name: taskID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.TaskStatusResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kshmirko_lidar-platform-go_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Poll async task status
      tags:
      - tasks
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Example: "Bearer eyJhbGciOiJIUzI1NiIs..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"